requirements.txt
asgiref==3.4.1
Django==3.2.5
django-cors-headers==3.7.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
gitdb==4.0.7
GitPython==3.1.18
pytz==2021.1
smmap==4.0.0
sqlparse==0.4.1
